pyproject.toml
[tool.poetry]
name = "document_manager"
version = "0.1.0"
description = "Document Manager"
authors = ["Rodrigo Saraiva <rodrigobezerrasaraiva@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
django-extensions = "^3.2.0"
drf-spectacular = "^0.22.1"
django-filter = "^22.1"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"

[tool.flake8]
max-line-length = 91
extend-ignore = "E501"

[tool.isort]
profile = "black"
line_length = 79
force_sort_within_sections = "False"
order_by_type = "False"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
